+++ src/lib.rs
@@ -118,7 +118,7 @@ pub fn dumps(
     } else {
         optsbits = 0
     };
-    match encode::serialize(obj.as_ptr(), pydef, optsbits as u8) {
+    match encode::serialize(obj.as_ptr(), pydef, optsbits as u16) {
         Ok(val) => unsafe { Ok(PyObject::from_owned_ptr(py, val.as_ptr())) },
         Err(err) => Err(err),
     }
